Reorganize Horreum files into core/ and subclass_methods/ directories
- Move core functionality (serialization, database_commands, utils, connection) to core/
- Move settings into shared/
- Move subclass methods (definition, management, related_fields_management) to subclass/
- Update require_relative paths
- Fix dependency order in familia.rb (features and data_type before horreum)
- Preserve git history using git mv

The key insight: focus on where the code runs (Horreum vs subclasses)
rather than instance/class distinction, since Ruby modules blur that line.
