+++ lib/familia/horreum/core/connection.rb
