+++ lib/familia/horreum/core/serialization.rb
