+++ lib/familia/horreum/shared/settings.rb
